+++ jobbergate-api/CHANGELOG.md
@@ -4,11 +4,14 @@ This file keeps track of all notable changes to jobbergate-api
 
 ## Unreleased
 
+- Added constraints to prevent long strings from being inserted into the database
+- Added constraints to prevent negative id numbers from being inserted into the database
 
 ## 4.2.0a6 -- 2023-12-12
 ## 4.2.0a5 -- 2023-12-12
 ## 4.2.0a4 -- 2023-12-11
 ## 4.2.0a3 -- 2023-11-30
+
 - Added support for on-site job submissions [ASP-4238]
 
 ## 4.2.0a2 -- 2023-11-29